requirements.txt
streamlit
Pillow
numpy
torch
torchvision
torchaudio
opencv-python
matplotlib
plotly